move util methods to package/utils.go, comments
